Initial python prototype, and webpack+ts
Wrote a proof of concept in python to prove that this would actually be pretty interesting
to explore. Just reads in the history file and does the easy ones with a simple dictionary.

On the webside: started a typescript project that builds to a single .js file that's loaded.
So I don't have to write vanilla JS, but can still compile everything to something that can
run locally.

Needed to follow several webpack tutorials to get it working:
* https://webpack.js.org/guides/getting-started/
* https://webpack.js.org/guides/typescript/